CHANGED: source code structure
This commit goes with source code structure changed.
