+++ Sources/PerseusGeoLocationKit/Models/LocationAuthorization.swift
@@ -1,0 +1,28 @@
+//
+//  LocationAuthorization.swift
+//  PerseusGeoLocationKit
+//
+//  Created by Mikhail Zhigulin in 7531.
+//
+//  Copyright © 7531 Mikhail Zhigulin of Novosibirsk.
+//
+//  Licensed under the MIT license. See LICENSE file.
+//  All rights reserved.
+//
+
+import Foundation
+
+public enum LocationAuthorization: CustomStringConvertible {
+
+    public var description: String {
+        switch self {
+        case .whenInUse:
+            return "When-in-use"
+        case .always:
+            return "Always"
+        }
+    }
+
+    case whenInUse
+    case always
+}
