+++ Sources/PerseusGeoLocationKit/Models/Result.swift
@@ -1,0 +1,18 @@
+//
+//  Result.swift
+//  PerseusGeoLocationKit
+//
+//  Created by Mikhail Zhigulin in 7531.
+//
+//  Copyright © 7531 Mikhail Zhigulin of Novosibirsk.
+//
+//  Licensed under the MIT license. See LICENSE file.
+//  All rights reserved.
+//
+
+import Foundation
+
+public enum Result<Value, Error: Swift.Error> {
+    case success(Value)
+    case failure(Error)
+}
